requirements.txt
# DermAI Pro - Professional Dermatology AI System
# Requirements for Python 3.8+

# Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
requests>=2.25.0

# GUI Framework
customtkinter>=5.0.0
tkinter-tooltip>=2.0.0

# Image Processing
scikit-image>=0.18.0
matplotlib>=3.3.0
scipy>=1.7.0

# Machine Learning Support
scikit-learn>=1.0.0
joblib>=1.0.0

# Logging and Configuration
colorlog>=6.0.0
pyyaml>=5.4.0
configparser>=5.0.0

# Medical Data Processing
pandas>=1.3.0
jsonschema>=3.2.0

# Ollama Integration
ollama>=0.1.0

# Optional: Advanced Image Processing
# mediapipe>=0.8.0  # Uncomment if MediaPipe is available

# Development and Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
